Move core bundle source into src directory
