+++ modules/core/bundle/index.ts
@@ -1,14 +1,14 @@
 // We use `require` here because luma and deck core must be imported before `global`
-import * as LumaGL from './lumagl';
-import {registerLoaders, load, parse, fetchFile} from '@loaders.gl/core';
+import * as LumaGL from '../src/scripting/lumagl';
+import * as LoadersGL from '../src/scripting/loadersgl';
 
 globalThis.luma = globalThis.luma || {};
 globalThis.loaders = globalThis.loaders || {};
 
 Object.assign(globalThis.luma, LumaGL);
-Object.assign(globalThis.loaders, {registerLoaders, load, parse, fetchFile});
+Object.assign(globalThis.loaders, LoadersGL);
 
 export * from '../src';
 export {register as _registerLoggers} from '../src/debug';
 
-export {default as DeckGL} from './deckgl';
+export {default as DeckGL} from '../src/scripting/deckgl';
